Update to MicroProfile 2.0.0-RC3
