+++ parent/pom.xml
@@ -164,9 +164,9 @@
         <cxf.lucene.version>8.2.0</cxf.lucene.version>
         <cxf.maven.core.version>3.6.3</cxf.maven.core.version>
         <cxf.micrometer.version>1.6.1</cxf.micrometer.version>
-        <cxf.microprofile.config.version>1.2</cxf.microprofile.config.version>
-        <cxf.microprofile.rest.client.version>2.0-RC2</cxf.microprofile.rest.client.version>
-        <cxf.microprofile.openapi.version>1.1.2</cxf.microprofile.openapi.version>        
+        <cxf.microprofile.config.version>2.0</cxf.microprofile.config.version>
+        <cxf.microprofile.rest.client.version>2.0-RC3</cxf.microprofile.rest.client.version>
+        <cxf.microprofile.openapi.version>2.0-RC3</cxf.microprofile.openapi.version>        
         <cxf.mina.version>2.1.3</cxf.mina.version>
         <cxf.mockito.version>3.4.6</cxf.mockito.version>
         <cxf.msv.version>2013.6.1</cxf.msv.version>
