+++ systests/microprofile/client/weld/testng.xml
@@ -4,11 +4,11 @@
     <test name="All TCK Tests">
         <packages>
             <package name="org.eclipse.microprofile.rest.client.tck" />
-            <package name="org.eclipse.microprofile.rest.client.tck.asynctests" />
             <package name="org.eclipse.microprofile.rest.client.tck.cditests" />
-            <package name="org.eclipse.microprofile.rest.client.tck.sse" />
+            <package name="org.eclipse.microprofile.rest.client.tck.asynctests" />
+            <package name="org.eclipse.microprofile.rest.client.tck.sse" /> 
             <package name="org.eclipse.microprofile.rest.client.tck.ssl" />
-            <package name="org.eclipse.microprofile.rest.client.tck.timeout" />
+            <package name="org.eclipse.microprofile.rest.client.tck.timeout" /> 
         </packages>
     </test>
 </suite>
